remove a11y and appearance middleman screen
