+++ bskyweb/cmd/bskyweb/server.go
@@ -256,7 +256,6 @@ func serve(cctx *cli.Context) error {
 	e.GET("/settings/account", server.WebGeneric)
 	e.GET("/settings/privacy-and-security", server.WebGeneric)
 	e.GET("/settings/content-and-media", server.WebGeneric)
-	e.GET("/settings/accessibility-and-appearance", server.WebGeneric)
 	e.GET("/settings/about", server.WebGeneric)
 	e.GET("/sys/debug", server.WebGeneric)
 	e.GET("/sys/debug-mod", server.WebGeneric)
